% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_onnxruntime}
\alias{install_onnxruntime}
\title{Install onnxruntime python package}
\usage{
install_onnxruntime(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  version = "1.23.2",
  envname = NULL,
  restart_session = TRUE,
  conda_python_version = NULL,
  ...,
  pip_ignore_installed = TRUE,
  python_version = conda_python_version
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{version}{onnxruntime version to install.}

\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio)}

\item{...}{other arguments passed to \code{\link[reticulate:conda-tools]{reticulate::conda_install()}} or
\code{\link[reticulate:virtualenv-tools]{reticulate::virtualenv_install()}}, depending on the \code{method} used.}

\item{pip_ignore_installed}{Whether pip should ignore installed python
packages and reinstall all already installed python packages. This defaults
to \code{TRUE}, to ensure that onnxruntime dependencies like NumPy are compatible
with the prebuilt onnxruntime binaries.}

\item{python_version, conda_python_version}{Pass a string like "3.8" to
request that conda install a specific Python version. This is ignored when
attempting to install in a Python virtual environment.}
}
\description{
\code{install_onnxruntime()} installs just the onnxruntime python package and it's
direct dependencies to inference model.
}
