% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kordivider-package.R
\docType{package}
\name{kordivider-package}
\alias{kordivider}
\alias{kordivider-package}
\title{kordivider: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/kordivider}
  \item Report bugs at \url{https://github.com/mrchypark/kordivider/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
