% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theeuh-package.R
\docType{package}
\name{theeuh-package}
\alias{theeuh}
\alias{theeuh-package}
\title{theeuh: Write Spaces to Korean Sentences}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Restores spacing in Korean sentences using a bundled neural network model and 'ONNX Runtime' \url{https://onnxruntime.ai/} through 'churon'. Inference runs locally without a web service or automatic downloads.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/theeuh}
  \item \url{https://mrchypark.github.io/theeuh/}
  \item Report bugs at \url{https://github.com/mrchypark/theeuh/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

Authors:
\itemize{
  \item Heewon Jeon (Author of KoSpacing code and bundled model assets)
}

}
\keyword{internal}
