% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theeuh-package.R
\docType{package}
\name{theeuh-package}
\alias{theeuh}
\alias{theeuh-package}
\title{theeuh: Add Spaces To Korean Sentences}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a function to add spaces to Korean sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/kordivider}
  \item \url{https://mrchypark.github.io/kordivider/}
  \item Report bugs at \url{https://github.com/mrchypark/kordivider/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
